# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuquet Automa Core Bridge API",
    "description": "Lightweight native process launcher, WebSocket event hub, and SQLite storage bridge for Automa WebExtension.",
    "contact": {
      "name": "Tuquet Ecosystem",
      "url": "https://github.com/tuquet/tuquet-automa"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "Local Daemon Bridge Server"
    }
  ],
  "paths": {
...
    "/api/v1/storage/workflows": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "List all workflows from storage database",
        "description": "Retrieves all workflow items persisted in the central SQLite database.",
        "operationId": "get_storage_workflows",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of workflows to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip for pagination (default 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "search",
            "in": "query",
            "description": "Optional search query to filter workflows by name, ID or description",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workflows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowStorageItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database read error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Create or persist a workflow in database",
        "description": "Creates a new workflow record or updates an existing record in the SQLite database.",
        "operationId": "create_storage_workflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowStorageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workflow created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStorageItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid workflow payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database write error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/workflows/import": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Import a workflow JSON into storage database",
        "description": "Parses and imports a workflow JSON object (from file or client) into SQLite database.",
        "operationId": "import_storage_workflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportWorkflowStorageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workflow imported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStorageItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid workflow JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database write error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/workflows/{id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Get a workflow by ID from storage database",
        "description": "Retrieves the full AST and metadata of a workflow by its unique ID from SQLite database.",
        "operationId": "get_storage_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique workflow identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStorageItem"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database read error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Storage"
        ],
        "summary": "Update a workflow in storage database",
        "description": "Modifies the metadata or graph AST of an existing workflow in SQLite database.",
        "operationId": "update_storage_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique workflow identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowStorageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workflow updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStorageItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database write error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete a workflow from storage database",
        "description": "Removes a workflow record from the central SQLite database by ID.",
        "operationId": "delete_storage_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique workflow identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkflowResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database deletion error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "CreateWorkflowStorageRequest": {
        "type": "object",
        "description": "Request payload for creating a new workflow in SQLite database",
        "required": [
          "name",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Workflow graph AST (nodes, edges, settings)"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional workflow description"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional UI icon name"
          },
          "id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional custom identifier (auto-generated if omitted)"
          },
          "name": {
            "type": "string",
            "description": "Workflow display name"
          },
          "version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional version (defaults to 1.0.0)"
          }
        },
        "example": {
          "data": {
            "edges": [],
            "nodes": [
              {
                "id": "node_1",
                "label": "trigger",
                "type": "trigger"
              }
            ]
          },
          "description": "Searches Bing and extracts results",
          "icon": "search",
          "id": "wf_bing_search",
          "name": "Bing Search Automation",
          "version": "1.0.0"
        }
      },
      "DeleteWorkflowResponse": {
        "type": "object",
        "description": "Response returned after successfully deleting a workflow",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Confirmation message"
          },
          "success": {
            "type": "boolean",
            "description": "True if deletion succeeded"
          }
        },
        "example": {
          "message": "Workflow 'wf_1' deleted",
          "success": true
        }
      },
...
      "ImportWorkflowStorageRequest": {
        "type": "object",
        "description": "Request payload for importing a workflow into SQLite database",
        "required": [
          "workflow"
        ],
        "properties": {
          "id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional custom ID to assign"
          },
          "workflow": {
            "type": "object",
            "description": "Raw workflow JSON content"
          }
        },
        "example": {
          "workflow": {
            "edges": [],
            "name": "Imported Google Search",
            "nodes": []
          }
        }
      },
...
      "UpdateWorkflowStorageRequest": {
        "type": "object",
        "description": "Request payload for updating an existing workflow in SQLite database",
        "properties": {
          "data": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional updated workflow graph AST"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional updated description"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional updated icon"
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional updated name"
          },
          "version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional updated version"
          }
        },
        "example": {
          "data": {
            "edges": [],
            "nodes": [
              {
                "id": "node_1",
                "label": "trigger",
                "type": "trigger"
              }
            ]
          },
          "description": "Updated description",
          "icon": "search",
          "name": "Updated Bing Search Automation",
          "version": "1.1.0"
        }
      },
      "WorkflowStorageItem": {
        "type": "object",
        "description": "Workflow descriptor stored in central SQLite database",
        "required": [
          "id",
          "name",
          "data",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp (ISO 8601)"
          },
          "data": {
            "type": "object",
            "description": "Workflow graph AST (nodes, edges, settings)"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional workflow description"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional UI icon identifier"
          },
          "id": {
            "type": "string",
            "description": "Unique workflow identifier"
          },
          "name": {
            "type": "string",
            "description": "Workflow display name"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last modification timestamp (ISO 8601)"
          },
          "version": {
            "type": "string",
            "description": "Semantic version of the workflow"
          }
        },
        "example": {
          "createdAt": "2026-08-26T00:00:00Z",
          "data": {
            "edges": [],
            "nodes": [
              {
                "id": "node_1",
                "label": "trigger",
                "type": "trigger"
              }
            ]
          },
          "description": "Searches Bing and extracts results",
          "icon": "search",
          "id": "wf_bing_search",
          "name": "Bing Search Automation",
          "updatedAt": "2026-08-26T00:00:00Z",
          "version": "1.0.0"
        }
      }
    }
  },
...
}